Refactors UmiTemplateLibrary tests into test_umi.py
